requirements.txt
google-generativeai>=0.3.0
google-genai>=1.16.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
gspread>=5.0.0
moviepy>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0
pydub>=0.25.0
fugashi>=1.3.0
unidic-lite>=1.0.8
janome>=0.5.0
gtts>=2.3.0
google-cloud-texttospeech>=2.14.0
modal>=0.64.0
faster-whisper>=1.0.0
anthropic>=0.40.0
